+++ src/vs/workbench/contrib/terminal/browser/terminalTabbedView.ts
@@ -23,6 +23,8 @@ import { localize } from 'vs/nls';
 import { openContextMenu } from 'vs/workbench/contrib/terminal/browser/terminalContextMenu';
 import { TerminalStorageKeys } from 'vs/workbench/contrib/terminal/common/terminalStorageKeys';
 import { TerminalContextKeys } from 'vs/workbench/contrib/terminal/common/terminalContextKey';
+import { getInstanceHoverInfo } from 'vs/workbench/contrib/terminal/browser/terminalTooltip';
+import { IHoverService } from 'vs/workbench/services/hover/browser/hover';
 
 const $ = dom.$;
 
@@ -77,6 +79,7 @@ export class TerminalTabbedView extends Disposable {
 		@IMenuService menuService: IMenuService,
 		@IStorageService private readonly _storageService: IStorageService,
 		@IContextKeyService contextKeyService: IContextKeyService,
+		@IHoverService private readonly _hoverService: IHoverService,
 	) {
 		super();
 
@@ -480,7 +483,19 @@ export class TerminalTabbedView extends Disposable {
 	}
 
 	showHover(focus: boolean) {
-		this._tabList.showHover(focus);
+		if (this._shouldShowTabs()) {
+			this._tabList.showHover(focus);
+			return;
+		}
+		const instance = this._terminalGroupService.activeInstance;
+		if (!instance) {
+			return;
+		}
+		this._hoverService.showHover({
+			...getInstanceHoverInfo(instance),
+			target: this._terminalContainer,
+			trapFocus: focus
+		}, focus);
 	}
 
 	private _focus() {
