+++ src/vs/workbench/contrib/terminal/browser/terminalTooltip.ts
@@ -6,6 +6,25 @@
 import { localize } from 'vs/nls';
 import { ITerminalInstance } from 'vs/workbench/contrib/terminal/browser/terminal';
 import { TerminalCapability } from 'vs/platform/terminal/common/capabilities/capabilities';
+import { IHoverAction } from 'vs/workbench/services/hover/browser/hover';
+import { MarkdownString } from 'vs/base/common/htmlContent';
+
+export function getInstanceHoverInfo(instance: ITerminalInstance): { content: MarkdownString; actions: IHoverAction[] } {
+	let statusString = '';
+	const statuses = instance.statusList.statuses;
+	const actions = [];
+	for (const status of statuses) {
+		statusString += `\n\n---\n\n${status.icon ? `$(${status.icon?.id}) ` : ''}${status.tooltip || status.id}`;
+		if (status.hoverActions) {
+			actions.push(...status.hoverActions);
+		}
+	}
+
+	const shellIntegrationString = getShellIntegrationTooltip(instance, true);
+	const content = new MarkdownString(instance.title + shellIntegrationString + statusString, { supportThemeIcons: true });
+
+	return { content, actions };
+}
 
 export function getShellIntegrationTooltip(instance: ITerminalInstance, markdown: boolean): string {
 	const shellIntegrationCapabilities: TerminalCapability[] = [];
